+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed issue that prevented doodad sets from exporting/previewing under certain circumstances.
 - 3D model panning via W, A, S, D, Q, E keys implemented.
 - Allow users to manually configure the map texture split threshold in configuration.
+- Added 'Use Absolute Model Placement Paths' option.
 
 0.1.8 (11-01-2020)
 - NPC variant textures can now be selected/exported for creature models.